mb/ti/beaglebone: Rename beaglebone to beaglebone-black
Retargets the Beaglebone port to be a Beaglebone Black port. The
original Beaglebone (sometimes referred to as the Beaglebone White) has
been mostly superseded by the Beaglebone Black and is no longer
available for purchase.

This port currently is unfinished, so it makes sense to retarget it
towards available hardware. The original Beaglebone and Beaglebone Black
are very similar in any case (biggest material difference is 256MB of
RAM in the original vs. 512MB in the Black) so, if there were interest,
it probably would not be very difficult to get the port to work on the
original Beaglebone.

Change-Id: Ied0632badf19c20b37411375f0bce652453aef6e
Signed-off-by: Sam Lewis <sam.vr.lewis@gmail.com>